requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
PyMySQL==1.1.0
Flask-Migrate==4.0.5
cryptography==43.0.1
python-dotenv==1.0.0
requests